mid state of path searching

+++ LabyrinthOfShadows22/Assets/Scripts/Gameplay/Pathfinding/TileNode.cs
@@ -4,10 +4,10 @@ namespace Gameplay.Pathfinding
 {
     public class TileNode
     {
-        public float GCost { get; set; }
-        public float HCost { get; set; }
+        public int GCost { get; set; }
+        public int HCost { get; set; }
 
-        public float FCost => GCost + HCost;
+        public int FCost => GCost + HCost;
         
         public bool IsWalkable { get; set; }
         public Vector2Int Position { get; set; }
